--- src/Modules/add.js
@@ -1,3 +1,0 @@
-export default function addTask(arr, obj) {
-  return [...arr, obj];
-}
